Added Main Menu and options panel
Only exit button and Options button work so far.
Disable the following rows in the Windowhandler.cpp to bypass the main
menu 56 and 135 -137, and enable the following row 55
Modified the button class to have text alignment of Left, Center and
Right, the text is adjusted to the center of the button height. The
button class also has a std::function<void(void)> onClick that can take
custom functions. Small bugfix for windows not being openable after
closing.