Add code to list available mod files and list the selected mod file dependency
Also added OnOpen pure virtual on ModalBase to be required overload when the modal is opened, this allows for loading dynamic data and other things that should not be loaded in the constructor of the Modal itself, or data that needs to be refreshed between each open of the modal.

Also adds behavior to the NewFileModal to do all behavior it needs for now.
Remaining to do actually create the file and required metadata and save it to the file and then load it as the current file.