Add another menuitem to the editor main window, to save and reload dependencies.
The plan is to be able to reload dependencies if they are changed, e.g. a dependency resource mod that contains prefabs and whatnot to keep things separate, think like base dependency mod that needs to be updated.
It will also allow the user to have multiple sessions at the same time (at the cost of more memory used)